Extract StatsModuleListing component
